Move adapter tests into a separate test directory and split them up
